+++ docs/source/news.rst
@@ -6,7 +6,9 @@ What's New
 
 Features:
 
+* Eliot tasks can now more easily :ref:`span multiple threads <cross thread tasks>` using the new ``eliot.preserve_context`` API.
 * ``eliot-prettyprint`` command line tool now pretty prints field values in a more informative manner.
+
 
 0.10.1
 ^^^^^^